Former-commit-id: 7a878c2398691b592001a69fd90ffd65f974601c

+++ Programs/QuickEd/Classes/Utils/QtDavaConvertion.cpp
@@ -31,7 +31,7 @@ QString WideStringToQString(const DAVA::WideString& str)
 #ifdef __DAVAENGINE_MACOS__
     return QString::fromStdWString(str);
 #else
-    return QString((const QChar*)str.c_str(), str.length());
+    return QString((const QChar*)str.c_str(), static_cast<int>(str.length()));
 #endif
 }
